+++ main/managers/wifi_manager.c
@@ -2,6 +2,7 @@
 
 #include "managers/wifi_manager.h"
 #include "core/callbacks.h"  // For callback function declarations
+#include "core/ouis.h"       // For OUI vendor lookup
 #include "esp_crt_bundle.h"
 #include "esp_event.h"
 #include "esp_heap_caps.h" // Add include for heap stats
@@ -378,17 +379,6 @@ struct DeviceInfo {
     struct eth_addr mac;
 };
 
-typedef enum {
-    COMPANY_DLINK,
-    COMPANY_NETGEAR,
-    COMPANY_BELKIN,
-    COMPANY_TPLINK,
-    COMPANY_LINKSYS,
-    COMPANY_ASUS,
-    COMPANY_ACTIONTEC,
-    COMPANY_UNKNOWN
-} ECompany;
-
 void wifi_manager_set_manual_disconnect(bool disconnect) {
     manual_disconnect = disconnect;
 }
@@ -575,76 +565,7 @@ static void add_station_ap_pair(const uint8_t *station_mac, const uint8_t *ap_bs
     }
 }
 
-typedef struct {
-    const char* oui;
-    ECompany company;
-} oui_company_map_t;
-
-static const oui_company_map_t OUI_MAP[] = {
-    {"00055D", COMPANY_DLINK}, {"000D88", COMPANY_DLINK}, {"000F3D", COMPANY_DLINK}, {"001195", COMPANY_DLINK}, {"001346", COMPANY_DLINK}, {"0015E9", COMPANY_DLINK}, {"00179A", COMPANY_DLINK}, {"00195B", COMPANY_DLINK}, {"001B11", COMPANY_DLINK},
-    {"001CF0", COMPANY_DLINK}, {"001E58", COMPANY_DLINK}, {"002191", COMPANY_DLINK}, {"0022B0", COMPANY_DLINK}, {"002401", COMPANY_DLINK}, {"00265A", COMPANY_DLINK}, {"00AD24", COMPANY_DLINK}, {"04BAD6", COMPANY_DLINK}, {"085A11", COMPANY_DLINK},
-    {"0C0E76", COMPANY_DLINK}, {"0CB6D2", COMPANY_DLINK}, {"1062EB", COMPANY_DLINK}, {"10BEF5", COMPANY_DLINK}, {"14D64D", COMPANY_DLINK}, {"180F76", COMPANY_DLINK}, {"1C5F2B", COMPANY_DLINK}, {"1C7EE5", COMPANY_DLINK}, {"1CAFF7", COMPANY_DLINK},
-    {"1CBDB9", COMPANY_DLINK}, {"283B82", COMPANY_DLINK}, {"302303", COMPANY_DLINK}, {"340804", COMPANY_DLINK}, {"340A33", COMPANY_DLINK}, {"3C1E04", COMPANY_DLINK}, {"3C3332", COMPANY_DLINK}, {"4086CB", COMPANY_DLINK}, {"409BCD", COMPANY_DLINK},
-    {"54B80A", COMPANY_DLINK}, {"5CD998", COMPANY_DLINK}, {"60634C", COMPANY_DLINK}, {"642943", COMPANY_DLINK}, {"6C198F", COMPANY_DLINK}, {"6C7220", COMPANY_DLINK}, {"744401", COMPANY_DLINK}, {"74DADA", COMPANY_DLINK}, {"78321B", COMPANY_DLINK},
-    {"78542E", COMPANY_DLINK}, {"7898E8", COMPANY_DLINK}, {"802689", COMPANY_DLINK}, {"84C9B2", COMPANY_DLINK}, {"8876B9", COMPANY_DLINK}, {"908D78", COMPANY_DLINK}, {"9094E4", COMPANY_DLINK}, {"9CD643", COMPANY_DLINK}, {"A06391", COMPANY_DLINK},
-    {"A0AB1B", COMPANY_DLINK}, {"A42A95", COMPANY_DLINK}, {"A8637D", COMPANY_DLINK}, {"ACF1DF", COMPANY_DLINK}, {"B437D8", COMPANY_DLINK}, {"B8A386", COMPANY_DLINK}, {"BC0F9A", COMPANY_DLINK}, {"BC2228", COMPANY_DLINK}, {"BCF685", COMPANY_DLINK},
-    {"C0A0BB", COMPANY_DLINK}, {"C4A81D", COMPANY_DLINK}, {"C4E90A", COMPANY_DLINK}, {"C8787D", COMPANY_DLINK}, {"C8BE19", COMPANY_DLINK}, {"C8D3A3", COMPANY_DLINK}, {"CCB255", COMPANY_DLINK}, {"D8FEE3", COMPANY_DLINK}, {"DCEAE7", COMPANY_DLINK},
-    {"E01CFC", COMPANY_DLINK}, {"E46F13", COMPANY_DLINK}, {"E8CC18", COMPANY_DLINK}, {"EC2280", COMPANY_DLINK}, {"ECADE0", COMPANY_DLINK}, {"F07D68", COMPANY_DLINK}, {"F0B4D2", COMPANY_DLINK}, {"F48CEB", COMPANY_DLINK}, {"F8E903", COMPANY_DLINK},
-    {"FC7516", COMPANY_DLINK},
-    {"00095B", COMPANY_NETGEAR}, {"000FB5", COMPANY_NETGEAR}, {"00146C", COMPANY_NETGEAR}, {"001B2F", COMPANY_NETGEAR}, {"001E2A", COMPANY_NETGEAR}, {"001F33", COMPANY_NETGEAR}, {"00223F", COMPANY_NETGEAR}, {"00224B", COMPANY_NETGEAR}, {"0026F2", COMPANY_NETGEAR},
-    {"008EF2", COMPANY_NETGEAR}, {"08028E", COMPANY_NETGEAR}, {"0836C9", COMPANY_NETGEAR}, {"08BD43", COMPANY_NETGEAR}, {"100C6B", COMPANY_NETGEAR}, {"100D7F", COMPANY_NETGEAR}, {"10DA43", COMPANY_NETGEAR}, {"1459C0", COMPANY_NETGEAR},  {"204E7F", COMPANY_NETGEAR},
-    {"20E52A", COMPANY_NETGEAR}, {"288088", COMPANY_NETGEAR}, {"289401", COMPANY_NETGEAR}, {"28C68E", COMPANY_NETGEAR}, {"2C3033", COMPANY_NETGEAR}, {"2CB05D", COMPANY_NETGEAR}, {"30469A", COMPANY_NETGEAR}, {"3498B5", COMPANY_NETGEAR},  {"3894ED", COMPANY_NETGEAR},
-    {"3C3786", COMPANY_NETGEAR}, {"405D82", COMPANY_NETGEAR}, {"44A56E", COMPANY_NETGEAR}, {"4C60DE", COMPANY_NETGEAR}, {"504A6E", COMPANY_NETGEAR}, {"506A03", COMPANY_NETGEAR}, {"54077D", COMPANY_NETGEAR}, {"58EF68", COMPANY_NETGEAR},  {"6038E0", COMPANY_NETGEAR},
-    {"6CB0CE", COMPANY_NETGEAR}, {"6CCDD6", COMPANY_NETGEAR}, {"744401", COMPANY_NETGEAR}, {"803773", COMPANY_NETGEAR}, {"841B5E", COMPANY_NETGEAR}, {"8C3BAD", COMPANY_NETGEAR}, {"941865", COMPANY_NETGEAR}, {"9C3DCF", COMPANY_NETGEAR},  {"9CC9EB", COMPANY_NETGEAR},
-    {"9CD36D", COMPANY_NETGEAR}, {"A00460", COMPANY_NETGEAR}, {"A021B7", COMPANY_NETGEAR}, {"A040A0", COMPANY_NETGEAR}, {"A42B8C", COMPANY_NETGEAR}, {"B03956", COMPANY_NETGEAR}, {"B07FB9", COMPANY_NETGEAR}, {"B0B98A", COMPANY_NETGEAR},  {"BCA511", COMPANY_NETGEAR},
-    {"C03F0E", COMPANY_NETGEAR}, {"C0FFD4", COMPANY_NETGEAR}, {"C40415", COMPANY_NETGEAR}, {"C43DC7", COMPANY_NETGEAR}, {"C89E43", COMPANY_NETGEAR}, {"CC40D0", COMPANY_NETGEAR}, {"DCEF09", COMPANY_NETGEAR}, {"E0469A", COMPANY_NETGEAR},  {"E046EE", COMPANY_NETGEAR},
-    {"E091F5", COMPANY_NETGEAR}, {"E4F4C6", COMPANY_NETGEAR}, {"E8FCAF", COMPANY_NETGEAR}, {"F87394", COMPANY_NETGEAR},
-    {"001150", COMPANY_BELKIN}, {"00173F", COMPANY_BELKIN}, {"0030BD", COMPANY_BELKIN}, {"08BD43", COMPANY_BELKIN}, {"149182", COMPANY_BELKIN}, {"24F5A2", COMPANY_BELKIN},
-    {"302303", COMPANY_BELKIN}, {"80691A", COMPANY_BELKIN}, {"94103E", COMPANY_BELKIN}, {"944452", COMPANY_BELKIN}, {"B4750E", COMPANY_BELKIN}, {"C05627", COMPANY_BELKIN},
-    {"C4411E", COMPANY_BELKIN}, {"D8EC5E", COMPANY_BELKIN}, {"E89F80", COMPANY_BELKIN}, {"EC1A59", COMPANY_BELKIN}, {"EC2280", COMPANY_BELKIN},
-    {"003192", COMPANY_TPLINK}, {"005F67", COMPANY_TPLINK}, {"1027F5", COMPANY_TPLINK}, {"14EBB6", COMPANY_TPLINK}, {"1C61B4", COMPANY_TPLINK}, {"203626", COMPANY_TPLINK}, {"2887BA", COMPANY_TPLINK},
-    {"30DE4B", COMPANY_TPLINK}, {"3460F9", COMPANY_TPLINK}, {"3C52A1", COMPANY_TPLINK}, {"40ED00", COMPANY_TPLINK}, {"482254", COMPANY_TPLINK}, {"5091E3", COMPANY_TPLINK}, {"54AF97", COMPANY_TPLINK},
-    {"5C628B", COMPANY_TPLINK}, {"5CA6E6", COMPANY_TPLINK}, {"5CE931", COMPANY_TPLINK}, {"60A4B7", COMPANY_TPLINK}, {"687FF0", COMPANY_TPLINK}, {"6C5AB0", COMPANY_TPLINK}, {"788CB5", COMPANY_TPLINK},
-    {"7CC2C6", COMPANY_TPLINK}, {"9C5322", COMPANY_TPLINK}, {"9CA2F4", COMPANY_TPLINK}, {"A842A1", COMPANY_TPLINK}, {"AC15A2", COMPANY_TPLINK}, {"B0A7B9", COMPANY_TPLINK}, {"B4B024", COMPANY_TPLINK},
-    {"C006C3", COMPANY_TPLINK}, {"CC68B6", COMPANY_TPLINK}, {"E848B8", COMPANY_TPLINK}, {"F0A731", COMPANY_TPLINK},
-    {"00045A", COMPANY_LINKSYS}, {"000625", COMPANY_LINKSYS}, {"000C41", COMPANY_LINKSYS}, {"000E08", COMPANY_LINKSYS}, {"000F66", COMPANY_LINKSYS}, {"001217", COMPANY_LINKSYS}, {"001310", COMPANY_LINKSYS}, {"0014BF", COMPANY_LINKSYS}, {"0016B6", COMPANY_LINKSYS},
-    {"001839", COMPANY_LINKSYS}, {"0018F8", COMPANY_LINKSYS}, {"001A70", COMPANY_LINKSYS}, {"001C10", COMPANY_LINKSYS}, {"001D7E", COMPANY_LINKSYS}, {"001EE5", COMPANY_LINKSYS}, {"002129", COMPANY_LINKSYS}, {"00226B", COMPANY_LINKSYS}, {"002369", COMPANY_LINKSYS},
-    {"00259C", COMPANY_LINKSYS}, {"002354", COMPANY_LINKSYS}, {"0024B2", COMPANY_LINKSYS}, {"003192", COMPANY_LINKSYS}, {"005F67", COMPANY_LINKSYS}, {"1027F5", COMPANY_LINKSYS}, {"14EBB6", COMPANY_LINKSYS}, {"1C61B4", COMPANY_LINKSYS}, {"203626", COMPANY_LINKSYS},
-    {"2887BA", COMPANY_LINKSYS}, {"305A3A", COMPANY_LINKSYS}, {"2CFDA1", COMPANY_LINKSYS}, {"302303", COMPANY_LINKSYS}, {"30469A", COMPANY_LINKSYS}, {"40ED00", COMPANY_LINKSYS}, {"482254", COMPANY_LINKSYS}, {"5091E3", COMPANY_LINKSYS}, {"54AF97", COMPANY_LINKSYS},
-    {"5CA2F4", COMPANY_LINKSYS}, {"5CA6E6", COMPANY_LINKSYS}, {"5CE931", COMPANY_LINKSYS}, {"60A4B7", COMPANY_LINKSYS}, {"687FF0", COMPANY_LINKSYS}, {"6C5AB0", COMPANY_LINKSYS}, {"788CB5", COMPANY_LINKSYS}, {"7CC2C6", COMPANY_LINKSYS}, {"9C5322", COMPANY_LINKSYS},
-    {"9CA2F4", COMPANY_LINKSYS}, {"A842A1", COMPANY_LINKSYS}, {"AC15A2", COMPANY_LINKSYS}, {"B0A7B9", COMPANY_LINKSYS}, {"B4B024", COMPANY_LINKSYS}, {"C006C3", COMPANY_LINKSYS}, {"CC68B6", COMPANY_LINKSYS}, {"E848B8", COMPANY_LINKSYS}, {"F0A731", COMPANY_LINKSYS},
-    {"000C6E", COMPANY_ASUS}, {"000EA6", COMPANY_ASUS}, {"00112F", COMPANY_ASUS}, {"0011D8", COMPANY_ASUS}, {"0013D4", COMPANY_ASUS}, {"0015F2", COMPANY_ASUS}, {"001731", COMPANY_ASUS}, {"0018F3", COMPANY_ASUS}, {"001A92", COMPANY_ASUS},
-    {"001BFC", COMPANY_ASUS}, {"001D60", COMPANY_ASUS}, {"001E8C", COMPANY_ASUS}, {"001FC6", COMPANY_ASUS}, {"002215", COMPANY_ASUS}, {"002354", COMPANY_ASUS}, {"00248C", COMPANY_ASUS}, {"002618", COMPANY_ASUS}, {"00E018", COMPANY_ASUS},
-    {"04421A", COMPANY_ASUS}, {"049226", COMPANY_ASUS}, {"04D4C4", COMPANY_ASUS}, {"04D9F5", COMPANY_ASUS}, {"08606E", COMPANY_ASUS}, {"086266", COMPANY_ASUS}, {"08BFB8", COMPANY_ASUS}, {"0C9D92", COMPANY_ASUS}, {"107B44", COMPANY_ASUS},
-    {"107C61", COMPANY_ASUS}, {"10BF48", COMPANY_ASUS}, {"10C37B", COMPANY_ASUS}, {"14DAE9", COMPANY_ASUS}, {"14DDA9", COMPANY_ASUS}, {"1831BF", COMPANY_ASUS}, {"1C872C", COMPANY_ASUS}, {"1CB72C", COMPANY_ASUS}, {"20CF30", COMPANY_ASUS},
-    {"244BFE", COMPANY_ASUS}, {"2C4D54", COMPANY_ASUS}, {"2C56DC", COMPANY_ASUS}, {"2CFDA1", COMPANY_ASUS}, {"305A3A", COMPANY_ASUS}, {"3085A9", COMPANY_ASUS}, {"3497F6", COMPANY_ASUS}, {"382C4A", COMPANY_ASUS}, {"38D547", COMPANY_ASUS},
-    {"3C7C3F", COMPANY_ASUS}, {"40167E", COMPANY_ASUS}, {"40B076", COMPANY_ASUS}, {"485B39", COMPANY_ASUS}, {"4CEDFB", COMPANY_ASUS}, {"50465D", COMPANY_ASUS}, {"50EBF6", COMPANY_ASUS}, {"5404A6", COMPANY_ASUS}, {"54A050", COMPANY_ASUS},
-    {"581122", COMPANY_ASUS}, {"6045CB", COMPANY_ASUS}, {"60A44C", COMPANY_ASUS}, {"60CF84", COMPANY_ASUS}, {"704D7B", COMPANY_ASUS}, {"708BCD", COMPANY_ASUS}, {"74D02B", COMPANY_ASUS}, {"7824AF", COMPANY_ASUS}, {"7C10C9", COMPANY_ASUS},
-    {"88D7F6", COMPANY_ASUS}, {"90E6BA", COMPANY_ASUS}, {"9C5C8E", COMPANY_ASUS}, {"A036BC", COMPANY_ASUS}, {"A85E45", COMPANY_ASUS}, {"AC220B", COMPANY_ASUS}, {"AC9E17", COMPANY_ASUS}, {"B06EBF", COMPANY_ASUS}, {"BCAEC5", COMPANY_ASUS},
-    {"BCEE7B", COMPANY_ASUS}, {"C86000", COMPANY_ASUS}, {"C87F54", COMPANY_ASUS}, {"CC28AA", COMPANY_ASUS}, {"D017C2", COMPANY_ASUS}, {"D45D64", COMPANY_ASUS}, {"D850E6", COMPANY_ASUS}, {"E03F49", COMPANY_ASUS}, {"E0CB4E", COMPANY_ASUS},
-    {"E89C25", COMPANY_ASUS}, {"F02F74", COMPANY_ASUS}, {"F07959", COMPANY_ASUS}, {"F46D04", COMPANY_ASUS}, {"F832E4", COMPANY_ASUS}, {"FC3497", COMPANY_ASUS}, {"FCC233", COMPANY_ASUS},
-    {"000FB3", COMPANY_ACTIONTEC}, {"001505", COMPANY_ACTIONTEC}, {"001801", COMPANY_ACTIONTEC}, {"001EA7", COMPANY_ACTIONTEC}, {"001F90", COMPANY_ACTIONTEC}, {"0020E0", COMPANY_ACTIONTEC},
-    {"00247B", COMPANY_ACTIONTEC}, {"002662", COMPANY_ACTIONTEC}, {"0026B8", COMPANY_ACTIONTEC}, {"007F28", COMPANY_ACTIONTEC}, {"0C6127", COMPANY_ACTIONTEC}, {"105F06", COMPANY_ACTIONTEC},
-    {"10785B", COMPANY_ACTIONTEC}, {"109FA9", COMPANY_ACTIONTEC}, {"181BEB", COMPANY_ACTIONTEC}, {"207600", COMPANY_ACTIONTEC}, {"408B07", COMPANY_ACTIONTEC}, {"4C8B30", COMPANY_ACTIONTEC},
-    {"5C35FC", COMPANY_ACTIONTEC}, {"7058A4", COMPANY_ACTIONTEC}, {"70F196", COMPANY_ACTIONTEC}, {"70F220", COMPANY_ACTIONTEC}, {"84E892", COMPANY_ACTIONTEC}, {"941C56", COMPANY_ACTIONTEC},
-    {"9C1E95", COMPANY_ACTIONTEC}, {"A0A3E2", COMPANY_ACTIONTEC}, {"A83944", COMPANY_ACTIONTEC}, {"E86FF2", COMPANY_ACTIONTEC}, {"F8E4FB", COMPANY_ACTIONTEC}, {"FC2BB2", COMPANY_ACTIONTEC},
-};
-
-// Function to match the BSSID to a company based on OUI
-ECompany match_bssid_to_company(const uint8_t *bssid) {
-    char oui[7];
-    snprintf(oui, sizeof(oui), "%02X%02X%02X", bssid[0], bssid[1], bssid[2]);
-
-    for (size_t i = 0; i < sizeof(OUI_MAP) / sizeof(OUI_MAP[0]); i++) {
-        if (strcmp(oui, OUI_MAP[i].oui) == 0) {
-            return OUI_MAP[i].company;
-        }
-    }
-
-    return COMPANY_UNKNOWN;
-}
-
-// Helper macro to check for broadcast/multicast addresses
+// helper macro to check for broadcast/multicast addresses
 #define IS_BROADCAST_OR_MULTICAST(addr) (((addr)[0] & 0x01) || (memcmp((addr), "\xff\xff\xff\xff\xff\xff", 6) == 0))
 
 // Function to check if a station MAC already exists in the list
@@ -763,11 +684,35 @@ void wifi_stations_sniffer_callback(void *buf, wifi_promiscuous_pkt_type_t type)
              strcpy(ssid_str, "(Hidden)");
         }
 
-        glog(
-            "New Station: %02X:%02X:%02X:%02X:%02X:%02X -> Associated AP: %s (%02X:%02X:%02X:%02X:%02X:%02X)\n",
-            station_mac[0], station_mac[1], station_mac[2], station_mac[3], station_mac[4], station_mac[5],
-            ssid_str, // Use SSID here
-            ap_bssid[0], ap_bssid[1], ap_bssid[2], ap_bssid[3], ap_bssid[4], ap_bssid[5]); // Use original ap_bssid
+        char station_mac_str[18];
+        snprintf(station_mac_str, sizeof(station_mac_str),
+                 "%02X:%02X:%02X:%02X:%02X:%02X",
+                 station_mac[0], station_mac[1], station_mac[2],
+                 station_mac[3], station_mac[4], station_mac[5]);
+
+        char ap_mac_str[18];
+        snprintf(ap_mac_str, sizeof(ap_mac_str),
+                 "%02X:%02X:%02X:%02X:%02X:%02X",
+                 ap_bssid[0], ap_bssid[1], ap_bssid[2],
+                 ap_bssid[3], ap_bssid[4], ap_bssid[5]);
+
+        char station_vendor[64] = "Unknown";
+        (void)ouis_lookup_vendor(station_mac_str, station_vendor, sizeof(station_vendor));
+
+        char ap_vendor[64] = "Unknown";
+        (void)ouis_lookup_vendor(ap_mac_str, ap_vendor, sizeof(ap_vendor));
+
+        glog("New Station:\n"
+             "     STA: %s\n"
+             "     STA Vendor: %s\n"
+             "     Associated AP: %s\n"
+             "     AP BSSID: %s\n"
+             "     AP Vendor: %s\n",
+             station_mac_str,
+             station_vendor,
+             ssid_str,
+             ap_mac_str,
+             ap_vendor);
 
         // Add the station and the *specific AP BSSID* it was seen with to the list
         add_station_ap_pair(station_mac, ap_bssid);
@@ -1918,9 +1863,10 @@ void wifi_manager_list_stations() {
         printf("No stations found.\n");
         return;
     }
-    printf("Listing all stations and their associated APs:\n");
+    printf("--- Station List (%d entries) ---\n", station_count);
+    TERMINAL_VIEW_ADD_TEXT("--- Station List (%d entries) ---\n", station_count);
     for (int i = 0; i < station_count; i++) {
-        char sanitized_ssid[33]; // Buffer for sanitized SSID
+        char sanitized_ssid[33];
         bool found = false;
         for (int j = 0; j < ap_count; j++) {
             if (memcmp(scanned_aps[j].bssid, station_ap_list[i].ap_bssid, 6) == 0) {
@@ -1931,17 +1877,39 @@ void wifi_manager_list_stations() {
         }
         if (!found) {
             strcpy(sanitized_ssid, "(Unknown AP)");
-         }
-        printf("[%d] Station MAC: %02X:%02X:%02X:%02X:%02X:%02X, AP SSID: %s, AP BSSID: %02X:%02X:%02X:%02X:%02X:%02X\n",
-               i,
-               station_ap_list[i].station_mac[0], station_ap_list[i].station_mac[1], station_ap_list[i].station_mac[2], station_ap_list[i].station_mac[3], station_ap_list[i].station_mac[4], station_ap_list[i].station_mac[5],
-               sanitized_ssid,
-               station_ap_list[i].ap_bssid[0], station_ap_list[i].ap_bssid[1], station_ap_list[i].ap_bssid[2], station_ap_list[i].ap_bssid[3], station_ap_list[i].ap_bssid[4], station_ap_list[i].ap_bssid[5]);
-        TERMINAL_VIEW_ADD_TEXT("[%d] Station MAC: %02X:%02X:%02X:%02X:%02X:%02X, AP SSID: %s, AP BSSID: %02X:%02X:%02X:%02X:%02X:%02X\n",
-                               i,
-                               station_ap_list[i].station_mac[0], station_ap_list[i].station_mac[1], station_ap_list[i].station_mac[2], station_ap_list[i].station_mac[3], station_ap_list[i].station_mac[4], station_ap_list[i].station_mac[5],
-                               sanitized_ssid,
-                               station_ap_list[i].ap_bssid[0], station_ap_list[i].ap_bssid[1], station_ap_list[i].ap_bssid[2], station_ap_list[i].ap_bssid[3], station_ap_list[i].ap_bssid[4], station_ap_list[i].ap_bssid[5]);
+        }
+
+        char sta_mac_str[18];
+        snprintf(sta_mac_str, sizeof(sta_mac_str), "%02X:%02X:%02X:%02X:%02X:%02X",
+                 station_ap_list[i].station_mac[0], station_ap_list[i].station_mac[1],
+                 station_ap_list[i].station_mac[2], station_ap_list[i].station_mac[3],
+                 station_ap_list[i].station_mac[4], station_ap_list[i].station_mac[5]);
+        char sta_vendor[64] = "Unknown";
+        if (!ouis_lookup_vendor(sta_mac_str, sta_vendor, sizeof(sta_vendor))) {
+            strncpy(sta_vendor, "Unknown", sizeof(sta_vendor) - 1);
+        }
+
+        char ap_mac_str[18];
+        snprintf(ap_mac_str, sizeof(ap_mac_str), "%02X:%02X:%02X:%02X:%02X:%02X",
+                 station_ap_list[i].ap_bssid[0], station_ap_list[i].ap_bssid[1],
+                 station_ap_list[i].ap_bssid[2], station_ap_list[i].ap_bssid[3],
+                 station_ap_list[i].ap_bssid[4], station_ap_list[i].ap_bssid[5]);
+        char ap_vendor[64] = "Unknown";
+        if (!ouis_lookup_vendor(ap_mac_str, ap_vendor, sizeof(ap_vendor))) {
+            strncpy(ap_vendor, "Unknown", sizeof(ap_vendor) - 1);
+        }
+
+        printf("[%d] Station MAC: %s\n", i, sta_mac_str);
+        printf("     Station Vendor: %s\n", sta_vendor);
+        printf("     Associated AP: %s\n", sanitized_ssid);
+        printf("     AP BSSID: %s\n", ap_mac_str);
+        printf("     AP Vendor: %s\n", ap_vendor);
+
+        TERMINAL_VIEW_ADD_TEXT("[%d] Station MAC: %s\n", i, sta_mac_str);
+        TERMINAL_VIEW_ADD_TEXT("     Station Vendor: %s\n", sta_vendor);
+        TERMINAL_VIEW_ADD_TEXT("     Associated AP: %s\n", sanitized_ssid);
+        TERMINAL_VIEW_ADD_TEXT("     AP BSSID: %s\n", ap_mac_str);
+        TERMINAL_VIEW_ADD_TEXT("     AP Vendor: %s\n", ap_vendor);
     }
 }
 
@@ -3844,18 +3812,13 @@ static void wifi_manager_print_ap_entry_formatted(uint16_t idx, const wifi_ap_re
     char sanitized_ssid[33];
     sanitize_ssid_and_check_hidden((uint8_t *)rec->ssid, sanitized_ssid, sizeof(sanitized_ssid));
 
-    ECompany company = match_bssid_to_company(rec->bssid);
-    const char *company_str = "Unknown";
-    switch (company) {
-        case COMPANY_DLINK: company_str = "DLink"; break;
-        case COMPANY_NETGEAR: company_str = "Netgear"; break;
-        case COMPANY_BELKIN: company_str = "Belkin"; break;
-        case COMPANY_TPLINK: company_str = "TPLink"; break;
-        case COMPANY_LINKSYS: company_str = "Linksys"; break;
-        case COMPANY_ASUS: company_str = "ASUS"; break;
-        case COMPANY_ACTIONTEC: company_str = "Actiontec"; break;
-        default: company_str = "Unknown"; break;
-    }
+    // lookup vendor using oui database
+    char mac_str[18];
+    snprintf(mac_str, sizeof(mac_str), "%02X:%02X:%02X:%02X:%02X:%02X",
+             rec->bssid[0], rec->bssid[1], rec->bssid[2],
+             rec->bssid[3], rec->bssid[4], rec->bssid[5]);
+    char vendor[64] = {0};
+    bool has_vendor = ouis_lookup_vendor(mac_str, vendor, sizeof(vendor));
 
     printf("[%u] SSID: %s,\n"
            "     BSSID: %02X:%02X:%02X:%02X:%02X:%02X,\n"
@@ -3908,9 +3871,9 @@ static void wifi_manager_print_ap_entry_formatted(uint16_t idx, const wifi_ap_re
     }
 #endif
 
-    if (strcmp(company_str, "Unknown") != 0) {
-        printf("     Company: %s\n", company_str);
-        TERMINAL_VIEW_ADD_TEXT("     Company: %s\n", company_str);
+    if (has_vendor) {
+        printf("     Vendor: %s\n", vendor);
+        TERMINAL_VIEW_ADD_TEXT("     Vendor: %s\n", vendor);
     }
 }
 void wifi_manager_print_scan_results_with_oui() {
@@ -3928,34 +3891,13 @@ void wifi_manager_print_scan_results_with_oui() {
         char sanitized_ssid[33];
         sanitize_ssid_and_check_hidden(scanned_aps[i].ssid, sanitized_ssid, sizeof(sanitized_ssid));
 
-        ECompany company = match_bssid_to_company(scanned_aps[i].bssid);
-        const char *company_str = "Unknown";
-        switch (company) {
-        case COMPANY_DLINK:
-            company_str = "DLink";
-            break;
-        case COMPANY_NETGEAR:
-            company_str = "Netgear";
-            break;
-        case COMPANY_BELKIN:
-            company_str = "Belkin";
-            break;
-        case COMPANY_TPLINK:
-            company_str = "TPLink";
-            break;
-        case COMPANY_LINKSYS:
-            company_str = "Linksys";
-            break;
-        case COMPANY_ASUS:
-            company_str = "ASUS";
-            break;
-        case COMPANY_ACTIONTEC:
-            company_str = "Actiontec";
-            break;
-        default:
-            company_str = "Unknown";
-            break;
-        }
+        // lookup vendor using oui database
+        char mac_str[18];
+        snprintf(mac_str, sizeof(mac_str), "%02X:%02X:%02X:%02X:%02X:%02X",
+                 scanned_aps[i].bssid[0], scanned_aps[i].bssid[1], scanned_aps[i].bssid[2],
+                 scanned_aps[i].bssid[3], scanned_aps[i].bssid[4], scanned_aps[i].bssid[5]);
+        char vendor[64] = {0};
+        bool has_vendor = ouis_lookup_vendor(mac_str, vendor, sizeof(vendor));
 
         glog("[%u] SSID: %s,\n"
              "     BSSID: %02X:%02X:%02X:%02X:%02X:%02X,\n"
@@ -4023,8 +3965,8 @@ void wifi_manager_print_scan_results_with_oui() {
             }
         }
 #endif
-        if (strcmp(company_str, "Unknown") != 0) {
-            glog("     Company: %s\n", company_str);
+        if (has_vendor) {
+            glog("     Vendor: %s\n", vendor);
         }
     }
 }
@@ -4816,17 +4758,14 @@ void wifi_manager_scanall_chart() {
         char sanitized_ssid[33];
         sanitize_ssid_and_check_hidden(scanned_aps[i].ssid, sanitized_ssid, sizeof(sanitized_ssid));
 
-        ECompany company = match_bssid_to_company(scanned_aps[i].bssid);
-        const char *company_str = "Unknown";
-        switch (company) {
-        case COMPANY_DLINK: company_str = "DLink"; break;
-        case COMPANY_NETGEAR: company_str = "Netgear"; break;
-        case COMPANY_BELKIN: company_str = "Belkin"; break;
-        case COMPANY_TPLINK: company_str = "TPLink"; break;
-        case COMPANY_LINKSYS: company_str = "Linksys"; break;
-        case COMPANY_ASUS: company_str = "ASUS"; break;
-        case COMPANY_ACTIONTEC: company_str = "Actiontec"; break;
-        default: company_str = "Unknown"; break;
+        // lookup vendor using oui database
+        char mac_str[18];
+        snprintf(mac_str, sizeof(mac_str), "%02X:%02X:%02X:%02X:%02X:%02X",
+                 scanned_aps[i].bssid[0], scanned_aps[i].bssid[1], scanned_aps[i].bssid[2],
+                 scanned_aps[i].bssid[3], scanned_aps[i].bssid[4], scanned_aps[i].bssid[5]);
+        char vendor[64] = {0};
+        if (!ouis_lookup_vendor(mac_str, vendor, sizeof(vendor))) {
+            strncpy(vendor, "Unknown", sizeof(vendor) - 1);
         }
 
         // Print AP details line
@@ -4834,7 +4773,7 @@ void wifi_manager_scanall_chart() {
         snprintf(ap_details_line, sizeof(ap_details_line), ap_format, sanitized_ssid,
                  scanned_aps[i].bssid[0], scanned_aps[i].bssid[1], scanned_aps[i].bssid[2],
                  scanned_aps[i].bssid[3], scanned_aps[i].bssid[4], scanned_aps[i].bssid[5],
-                 scanned_aps[i].primary, company_str);
+                 scanned_aps[i].primary, vendor);
         printf("%s\n", ap_details_line);
         TERMINAL_VIEW_ADD_TEXT("%s\n", ap_details_line);
 
@@ -4842,15 +4781,27 @@ void wifi_manager_scanall_chart() {
         // Find and print associated stations for this AP
         for (int j = 0; j < station_count; j++) {
             if (memcmp(station_ap_list[j].ap_bssid, scanned_aps[i].bssid, 6) == 0) {
-                // Print station MAC using the new format
-                char sta_details_line[100];
-                snprintf(sta_details_line, sizeof(sta_details_line), sta_format,
+                // lookup vendor for station mac
+                char sta_mac_str[18];
+                snprintf(sta_mac_str, sizeof(sta_mac_str), "%02X:%02X:%02X:%02X:%02X:%02X",
                          station_ap_list[j].station_mac[0], station_ap_list[j].station_mac[1],
                          station_ap_list[j].station_mac[2], station_ap_list[j].station_mac[3],
                          station_ap_list[j].station_mac[4], station_ap_list[j].station_mac[5]);
-                printf("%s\n", sta_details_line);
-                TERMINAL_VIEW_ADD_TEXT("%s\n", sta_details_line);
-                station_found_for_ap = true; // Mark that we printed at least one station
+                char sta_vendor[64] = {0};
+                bool has_sta_vendor = ouis_lookup_vendor(sta_mac_str, sta_vendor, sizeof(sta_vendor));
+
+                // Print station MAC using the new format
+                char sta_details_line[150];
+                if (has_sta_vendor) {
+                    snprintf(sta_details_line, sizeof(sta_details_line), "%s (%s)",
+                             sta_mac_str, sta_vendor);
+                } else {
+                    snprintf(sta_details_line, sizeof(sta_details_line), "%s",
+                             sta_mac_str);
+                }
+                printf("    STA: %s\n", sta_details_line);
+                TERMINAL_VIEW_ADD_TEXT("    STA: %s\n", sta_details_line);
+                station_found_for_ap = true;
             }
         }